VPN-4508: Allow custom DNS servers to be used.
