+++ nym-vpn-core/crates/nym-vpn-proto/src/conversions/mod.rs
@@ -15,6 +15,7 @@ pub mod vpn_api_client;
 pub mod vpnd;
 
 pub use error::ConversionError;
+use std::{net::IpAddr, str::FromStr};
 
 use crate::proto;
 
@@ -52,6 +53,39 @@ impl From<u8> for proto::Threshold {
     fn from(performance: u8) -> Self {
         Self {
             min_performance: performance.into(),
+        }
+    }
+}
+
+impl TryFrom<proto::CustomDns> for Option<Vec<IpAddr>> {
+    type Error = ConversionError;
+
+    fn try_from(value: proto::CustomDns) -> Result<Self, Self::Error> {
+        if value.ips.is_empty() {
+            Ok(None)
+        } else {
+            let ip_addrs: Vec<IpAddr> = value
+                .ips
+                .into_iter()
+                .map(|ip| {
+                    IpAddr::from_str(&ip)
+                        .map_err(|e| ConversionError::Generic(format!("Invalid IP address: {e}")))
+                })
+                .collect::<Result<Vec<_>, _>>()?;
+
+            Ok(Some(ip_addrs))
+        }
+    }
+}
+
+impl From<Option<Vec<IpAddr>>> for proto::CustomDns {
+    fn from(value: Option<Vec<IpAddr>>) -> Self {
+        match value {
+            Some(ips) => {
+                let ip_strings = ips.into_iter().map(|ip| ip.to_string()).collect();
+                proto::CustomDns { ips: ip_strings }
+            }
+            None => proto::CustomDns { ips: vec![] },
         }
     }
 }
