--- nym-vpn-core/crates/nym-vpnd/src/service/config.rs
@@ -1,1572 +1,0 @@
-// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: GPL-3.0-only
-
-use super::error::{Error, Result};
-use nym_common::trace_err_chain;
-use nym_registration_client::MixnetClientConfig;
-use nym_vpn_lib::{
-    DEFAULT_MIN_GATEWAY_PERFORMANCE, DEFAULT_MIN_MIXNODE_PERFORMANCE,
-    tunnel_state_machine::{
-        DnsOptions, GatewayPerformanceOptions, MixnetTunnelOptions, TunnelSettings,
-        WireguardMultihopMode, WireguardTunnelOptions,
-    },
-};
-use nym_vpn_lib_types::{
-    EntryPoint, ExitPoint, NodeIdentity, Recipient, TunnelEvent, TunnelType, VpnServiceConfig,
-};
-use serde::{Deserialize, Serialize, de::DeserializeOwned};
-use std::{
-    fmt,
-    net::IpAddr,
-    path::{Path, PathBuf},
-    str::FromStr,
-};
-use tokio::{
-    fs,
-    io::{self, AsyncWriteExt},
-    sync::broadcast,
-};
-
-#[cfg(unix)]
-use std::os::unix::fs::PermissionsExt;
-
-#[cfg(not(windows))]
-const DEFAULT_DATA_DIR: &str = "/var/lib/nym-vpnd";
-#[cfg(not(windows))]
-const DEFAULT_LOG_DIR: &str = "/var/log/nym-vpnd";
-#[cfg(not(windows))]
-const DEFAULT_CONFIG_DIR: &str = "/etc/nym";
-pub const DEFAULT_CONFIG_FILE_TOML: &str = "nym-vpnd.toml";
-pub const DEFAULT_CONFIG_FILE_JSON: &str = "nym-vpnd.json";
-pub const DEFAULT_LOG_FILE: &str = "nym-vpnd.log";
-pub const DEFAULT_OLD_LOG_FILE: &str = "nym-vpnd.old.log";
-
-pub const DEFAULT_GLOBAL_CONFIG_FILE_TOML: &str = "config.toml";
-pub const DEFAULT_GLOBAL_CONFIG_FILE_JSON: &str = "config.json";
-
-//
-// NetworkEnvironments
-//
-
-#[derive(Debug, Clone)]
-pub enum NetworkEnvironments {
-    Mainnet,
-    Sandbox,
-    Canary,
-    Evil,
-}
-
-impl fmt::Display for NetworkEnvironments {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            NetworkEnvironments::Mainnet => write!(f, "mainnet"),
-            NetworkEnvironments::Sandbox => write!(f, "sandbox"),
-            NetworkEnvironments::Canary => write!(f, "canary"),
-            NetworkEnvironments::Evil => write!(f, "evil"),
-        }
-    }
-}
-
-impl TryFrom<&str> for NetworkEnvironments {
-    type Error = &'static str;
-
-    fn try_from(env: &str) -> Result<Self, Self::Error> {
-        match env {
-            "mainnet" => Ok(NetworkEnvironments::Mainnet),
-            "sandbox" => Ok(NetworkEnvironments::Sandbox),
-            "canary" => Ok(NetworkEnvironments::Canary),
-            "evil" => Ok(NetworkEnvironments::Evil),
-            _ => Err("Invalid network environment"),
-        }
-    }
-}
-
-//
-// VpnServiceConfigManager
-//
-
-pub struct VpnServiceConfigManager {
-    json_config_path: PathBuf,
-    config: VpnServiceConfig,
-
-    // Used to send `ConfigChanged` events when the config is updated.
-    // It's only optional to simplify testing.
-    tunnel_event_tx: Option<broadcast::Sender<TunnelEvent>>,
-}
-
-impl VpnServiceConfigManager {
-    pub async fn new(
-        network_config_dir: &Path,
-        tunnel_event_tx: Option<broadcast::Sender<TunnelEvent>>,
-    ) -> Result<Self> {
-        let toml_config_path = network_config_dir.join(DEFAULT_CONFIG_FILE_TOML);
-        let json_config_path = network_config_dir.join(DEFAULT_CONFIG_FILE_JSON);
-        let (config, version) =
-            match Self::read_from_file(&toml_config_path, &json_config_path).await {
-                Ok((config, version)) => (config, version),
-                Err(e) => {
-                    trace_err_chain!(
-                        e,
-                        "Failed to read service config file {}; using default",
-                        json_config_path.display()
-                    );
-                    (VpnServiceConfig::default(), None)
-                }
-            };
-
-        let config_manager = Self {
-            json_config_path,
-            config,
-            tunnel_event_tx,
-        };
-
-        // If we didn't read the latest version then write the config straight back to file
-        if version != Some(VpnServiceConfigVersion::latest()) {
-            config_manager.write_to_file().await;
-        }
-
-        // If the deprecated TOML file exists then remove it
-        if toml_config_path.exists() {
-            tracing::info!(
-                "Removing deprecated config file {}",
-                toml_config_path.display()
-            );
-            if let Err(e) = fs::remove_file(&toml_config_path).await {
-                trace_err_chain!(e, "Failed to remove deprecated config file");
-            }
-        }
-
-        Ok(config_manager)
-    }
-
-    pub fn config(&self) -> &VpnServiceConfig {
-        &self.config
-    }
-
-    pub async fn set_config(&mut self, config: VpnServiceConfig) {
-        if self.config != config {
-            self.config = config;
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    pub async fn set_entry_point(&mut self, entry_point: EntryPoint) {
-        if self.config.entry_point != entry_point {
-            self.config.entry_point = entry_point;
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    pub async fn set_exit_point(&mut self, exit_point: ExitPoint) {
-        if self.config.exit_point != exit_point {
-            self.config.exit_point = exit_point;
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    #[allow(unused)]
-    pub async fn set_dns(&mut self, dns: Option<IpAddr>) {
-        if self.config.dns != dns {
-            self.config.dns = dns;
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    pub async fn set_disable_ipv6(&mut self, disable_ipv6: bool) {
-        if self.config.disable_ipv6 != disable_ipv6 {
-            self.config.disable_ipv6 = disable_ipv6;
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    pub async fn set_enable_two_hop(&mut self, enable_two_hop: bool) {
-        if self.config.enable_two_hop != enable_two_hop {
-            self.config.enable_two_hop = enable_two_hop;
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    pub async fn set_netstack(&mut self, netstack: bool) {
-        if self.config.netstack != netstack {
-            self.config.netstack = netstack;
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    pub async fn set_allow_lan(&mut self, allow_lan: bool) {
-        if self.config.allow_lan != allow_lan {
-            self.config.allow_lan = allow_lan;
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    pub async fn set_enable_bridges(&mut self, enable_bridges: bool) {
-        if self.config.enable_bridges != enable_bridges {
-            self.config.enable_bridges = enable_bridges;
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    pub async fn set_residential_exit(&mut self, residential_only: bool) {
-        if self.config.residential_exit != residential_only {
-            self.config.residential_exit = residential_only;
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    #[allow(unused)]
-    pub async fn set_disable_poisson_rate(&mut self, disable_poisson_rate: bool) {
-        if self.config.disable_poisson_rate != disable_poisson_rate {
-            self.config.disable_poisson_rate = disable_poisson_rate;
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    #[allow(unused)]
-    pub async fn set_disable_background_cover_traffic(&mut self, disable: bool) {
-        if self.config.disable_background_cover_traffic != disable {
-            self.config.disable_background_cover_traffic = disable;
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    #[allow(unused)]
-    pub async fn set_min_mixnode_performance(&mut self, min_mixnode_performance: Option<u8>) {
-        if self.config.min_mixnode_performance != min_mixnode_performance {
-            self.config.min_mixnode_performance = min_mixnode_performance.map(|u| u.min(100));
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    #[allow(unused)]
-    pub async fn set_min_gateway_mixnet_performance(
-        &mut self,
-        min_gateway_mixnet_performance: Option<u8>,
-    ) {
-        if self.config.min_gateway_mixnet_performance != min_gateway_mixnet_performance {
-            self.config.min_gateway_mixnet_performance =
-                min_gateway_mixnet_performance.map(|u| u.min(100));
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    #[allow(unused)]
-    pub async fn set_min_gateway_vpn_performance(
-        &mut self,
-        min_gateway_vpn_performance: Option<u8>,
-    ) {
-        if self.config.min_gateway_vpn_performance != min_gateway_vpn_performance {
-            self.config.min_gateway_vpn_performance =
-                min_gateway_vpn_performance.map(|u| u.min(100));
-            self.save_config_and_send_event().await;
-        }
-    }
-
-    async fn save_config_and_send_event(&self) {
-        // This function already logs
-        let _ = self.write_to_file().await;
-
-        // Notify all clients that the config has changed
-        if let Some(tx) = self.tunnel_event_tx.as_ref() {
-            match tx.send(TunnelEvent::ConfigChanged(Box::new(self.config.clone()))) {
-                Ok(recv_count) => {
-                    tracing::info!("Sent config changed event to {recv_count} receivers");
-                }
-                Err(e) => {
-                    tracing::error!("Failed to send config changed event: {e}");
-                }
-            }
-        }
-    }
-
-    /// Returns the configuration as well as the version read from file.
-    async fn read_from_file(
-        toml_config_path: &Path,
-        json_config_path: &Path,
-    ) -> Result<(VpnServiceConfig, Option<VpnServiceConfigVersion>)> {
-        let (config, version) = if json_config_path.exists() {
-            let ext_config = read_json_config_file::<VpnServiceConfigExt>(json_config_path)
-                .await
-                .map_err(Error::ConfigSetup)?;
-            let version = ext_config.version();
-
-            tracing::info!(
-                "Read service config version {version} from {}",
-                json_config_path.display()
-            );
-
-            let config = VpnServiceConfig::try_from(ext_config).map_err(Error::ConfigSetup)?;
-
-            (config, Some(version))
-        } else if toml_config_path.exists() {
-            let legacy_config = read_toml_config_file::<LegacyVpnServiceConfig>(toml_config_path)
-                .await
-                .map_err(Error::ConfigSetup)?;
-
-            tracing::info!("Read service config from {}", toml_config_path.display());
-
-            let config = VpnServiceConfig::try_from(legacy_config).map_err(Error::ConfigSetup)?;
-
-            (config, None)
-        } else {
-            tracing::info!("Using default service config");
-
-            (VpnServiceConfig::default(), None)
-        };
-
-        Ok((config, version))
-    }
-
-    async fn write_to_file(&self) -> bool {
-        let ext_config =
-            match VpnServiceConfigExt::try_from(&self.config).map_err(Error::ConfigSetup) {
-                Ok(ext_config) => ext_config,
-                Err(e) => {
-                    tracing::error!("Failed to convert service config to JSON: {e}");
-                    return false;
-                }
-            };
-        let version = ext_config.version();
-
-        match write_json_config_file(&self.json_config_path, &ext_config)
-            .await
-            .map_err(Error::ConfigSetup)
-        {
-            Ok(_) => {
-                tracing::info!(
-                    "Writing service config version {version} to {}",
-                    self.json_config_path.display()
-                );
-                true
-            }
-            Err(e) => {
-                tracing::error!(
-                    "Failed to write service config version {version} to {}: {e}",
-                    self.json_config_path.display()
-                );
-                false
-            }
-        }
-    }
-
-    pub fn generate_tunnel_settings(&self) -> TunnelSettings {
-        tracing::info!("Using config: {:?}", self.config);
-
-        let gateway_options = GatewayPerformanceOptions {
-            mixnet_min_performance: self.config.min_gateway_mixnet_performance,
-            vpn_min_performance: self.config.min_gateway_vpn_performance,
-        };
-
-        let mixnet_client_config = MixnetClientConfig {
-            disable_poisson_rate: self.config.disable_poisson_rate,
-            disable_background_cover_traffic: self.config.disable_background_cover_traffic,
-            min_mixnode_performance: Some(
-                self.config
-                    .min_mixnode_performance
-                    .unwrap_or(DEFAULT_MIN_MIXNODE_PERFORMANCE),
-            ),
-            min_gateway_performance: Some(
-                self.config
-                    .min_gateway_mixnet_performance
-                    .unwrap_or(DEFAULT_MIN_GATEWAY_PERFORMANCE),
-            ),
-        };
-
-        let tunnel_type = if self.config.enable_two_hop {
-            TunnelType::Wireguard
-        } else {
-            TunnelType::Mixnet
-        };
-
-        let dns = self
-            .config
-            .dns
-            .map(|addr| DnsOptions::Custom(vec![addr]))
-            .unwrap_or_default();
-
-        TunnelSettings {
-            enable_ipv6: !self.config.disable_ipv6,
-            allow_lan: self.config.allow_lan,
-            residential_exit: self.config.residential_exit,
-            tunnel_type,
-            mixnet_tunnel_options: MixnetTunnelOptions { mtu: None },
-            wireguard_tunnel_options: WireguardTunnelOptions {
-                multihop_mode: if self.config.netstack {
-                    WireguardMultihopMode::Netstack
-                } else {
-                    WireguardMultihopMode::TunTun
-                },
-                enable_bridges: self.config.enable_bridges,
-            },
-            gateway_performance_options: gateway_options,
-            mixnet_client_config: Some(mixnet_client_config),
-            entry_point: Box::new(self.config.entry_point.clone()),
-            exit_point: Box::new(self.config.exit_point.clone()),
-            dns,
-        }
-    }
-}
-
-//
-// External, versioned, representation of the vpn service config file.
-//
-
-type VpnServiceConfigExtLatest = VpnServiceConfigExtV2;
-
-/// Represents the version of the vpn service config file.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-enum VpnServiceConfigVersion {
-    V1,
-    V2,
-}
-
-impl VpnServiceConfigVersion {
-    /// Returns the latest version of the config file.
-    pub fn latest() -> Self {
-        VpnServiceConfigVersion::V2
-    }
-}
-
-impl std::fmt::Display for VpnServiceConfigVersion {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(match self {
-            VpnServiceConfigVersion::V1 => "v1",
-            VpnServiceConfigVersion::V2 => "v2",
-        })
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(tag = "version")]
-#[serde(rename_all = "snake_case")]
-enum VpnServiceConfigExt {
-    V1(VpnServiceConfigExtV1),
-    V2(VpnServiceConfigExtV2),
-}
-
-impl VpnServiceConfigExt {
-    fn version(&self) -> VpnServiceConfigVersion {
-        match self {
-            VpnServiceConfigExt::V1(_) => VpnServiceConfigVersion::V1,
-            VpnServiceConfigExt::V2(_) => VpnServiceConfigVersion::V2,
-        }
-    }
-}
-
-impl TryFrom<VpnServiceConfigExt> for VpnServiceConfig {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: VpnServiceConfigExt) -> Result<Self, Self::Error> {
-        match value {
-            VpnServiceConfigExt::V1(v1) => VpnServiceConfig::try_from(v1),
-            VpnServiceConfigExt::V2(v2) => VpnServiceConfig::try_from(v2),
-        }
-    }
-}
-
-impl TryFrom<&VpnServiceConfig> for VpnServiceConfigExt {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: &VpnServiceConfig) -> Result<Self, Self::Error> {
-        // Always construct the latest external representation
-        let latest = VpnServiceConfigExtLatest::try_from(value)?;
-        Ok(latest.into())
-    }
-}
-
-//
-// v1
-//
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-struct VpnServiceConfigExtV1 {
-    entry_point: EntryPointExtV1,
-    exit_point: ExitPointExtV1,
-}
-
-impl From<VpnServiceConfigExtV1> for VpnServiceConfigExt {
-    fn from(v1: VpnServiceConfigExtV1) -> Self {
-        VpnServiceConfigExt::V1(v1)
-    }
-}
-
-impl TryFrom<VpnServiceConfigExtV1> for VpnServiceConfig {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: VpnServiceConfigExtV1) -> Result<Self, Self::Error> {
-        let config = VpnServiceConfig {
-            entry_point: EntryPoint::try_from(value.entry_point)?,
-            exit_point: ExitPoint::try_from(value.exit_point)?,
-            ..Default::default()
-        };
-        Ok(config)
-    }
-}
-
-//
-// v2
-//
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-struct VpnServiceConfigExtV2 {
-    entry_point: EntryPointExtV2,
-    exit_point: ExitPointExtV2,
-    dns: Option<String>,
-    allow_lan: bool,
-    disable_ipv6: bool,
-    enable_two_hop: bool,
-    enable_bridges: bool,
-    netstack: bool,
-    disable_poisson_rate: bool,
-    disable_background_cover_traffic: bool,
-    min_mixnode_performance: Option<u8>,
-    min_gateway_mixnet_performance: Option<u8>,
-    min_gateway_vpn_performance: Option<u8>,
-    residential_exit: bool,
-}
-
-impl From<VpnServiceConfigExtV2> for VpnServiceConfigExt {
-    fn from(v2: VpnServiceConfigExtV2) -> Self {
-        VpnServiceConfigExt::V2(v2)
-    }
-}
-
-impl TryFrom<VpnServiceConfigExtV2> for VpnServiceConfig {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: VpnServiceConfigExtV2) -> Result<Self, Self::Error> {
-        let dns = value
-            .dns
-            .map(|addr| {
-                IpAddr::from_str(&addr)
-                    .map_err(|e| ConfigSetupError::IpAddress { error: Box::new(e) })
-            })
-            .transpose()?;
-
-        let config = VpnServiceConfig {
-            entry_point: EntryPoint::try_from(value.entry_point)?,
-            exit_point: ExitPoint::try_from(value.exit_point)?,
-            dns,
-            allow_lan: value.allow_lan,
-            disable_ipv6: value.disable_ipv6,
-            enable_two_hop: value.enable_two_hop,
-            enable_bridges: value.enable_bridges,
-            netstack: value.netstack,
-            disable_poisson_rate: value.disable_poisson_rate,
-            disable_background_cover_traffic: value.disable_background_cover_traffic,
-            min_mixnode_performance: value.min_mixnode_performance,
-            min_gateway_mixnet_performance: value.min_gateway_mixnet_performance,
-            min_gateway_vpn_performance: value.min_gateway_vpn_performance,
-            residential_exit: value.residential_exit,
-        };
-        Ok(config)
-    }
-}
-
-//
-// Latest (v2)
-//
-
-impl TryFrom<&VpnServiceConfig> for VpnServiceConfigExtLatest {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: &VpnServiceConfig) -> Result<Self, Self::Error> {
-        let ext_config = VpnServiceConfigExtLatest {
-            entry_point: EntryPointExtV2::try_from(&value.entry_point)?,
-            exit_point: ExitPointExtV2::try_from(&value.exit_point)?,
-            dns: value.dns.map(|addr| addr.to_string()),
-            allow_lan: value.allow_lan,
-            disable_ipv6: value.disable_ipv6,
-            enable_two_hop: value.enable_two_hop,
-            enable_bridges: value.enable_bridges,
-            netstack: value.netstack,
-            disable_poisson_rate: value.disable_poisson_rate,
-            disable_background_cover_traffic: value.disable_background_cover_traffic,
-            min_mixnode_performance: value.min_mixnode_performance,
-            min_gateway_mixnet_performance: value.min_gateway_mixnet_performance,
-            min_gateway_vpn_performance: value.min_gateway_vpn_performance,
-            residential_exit: value.residential_exit,
-        };
-        Ok(ext_config)
-    }
-}
-
-//
-// EntryPointExtV2
-//
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-enum EntryPointExtV2 {
-    Gateway { identity: String },
-    Country { two_letter_iso_country_code: String },
-    Region { region: String },
-    Random,
-}
-
-impl TryFrom<EntryPointExtV2> for EntryPoint {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: EntryPointExtV2) -> Result<Self, Self::Error> {
-        match value {
-            EntryPointExtV2::Gateway { ref identity } => EntryPoint::from_base58_string(identity)
-                .map_err(|e| ConfigSetupError::EntryPoint(e.to_string())),
-            EntryPointExtV2::Country {
-                two_letter_iso_country_code,
-            } => Ok(EntryPoint::Country {
-                two_letter_iso_country_code,
-            }),
-            EntryPointExtV2::Region { region } => Ok(EntryPoint::Region { region }),
-            EntryPointExtV2::Random => Ok(EntryPoint::Random),
-        }
-    }
-}
-
-impl TryFrom<&EntryPoint> for EntryPointExtV2 {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: &EntryPoint) -> Result<Self, Self::Error> {
-        match value {
-            EntryPoint::Gateway { identity } => Ok(EntryPointExtV2::Gateway {
-                identity: identity.to_base58_string(),
-            }),
-            EntryPoint::Country {
-                two_letter_iso_country_code,
-            } => Ok(EntryPointExtV2::Country {
-                two_letter_iso_country_code: two_letter_iso_country_code.clone(),
-            }),
-            EntryPoint::Region { region } => Ok(EntryPointExtV2::Region {
-                region: region.clone(),
-            }),
-            EntryPoint::Random => Ok(EntryPointExtV2::Random),
-        }
-    }
-}
-
-//
-// ExitPointExtV2
-//
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "snake_case")]
-enum ExitPointExtV2 {
-    Address { address: String },
-    Gateway { identity: String },
-    Country { two_letter_iso_country_code: String },
-    Region { region: String },
-    Random,
-}
-
-impl TryFrom<ExitPointExtV2> for ExitPoint {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: ExitPointExtV2) -> Result<Self, Self::Error> {
-        match value {
-            ExitPointExtV2::Address { address } => {
-                let recipient = Recipient::try_from_base58_string(&address)
-                    .map_err(|e| ConfigSetupError::ExitPoint(e.to_string()))?;
-                Ok(ExitPoint::Address {
-                    address: Box::new(recipient),
-                })
-            }
-            ExitPointExtV2::Gateway { identity } => {
-                let node_identity = nym_vpn_lib_types::NodeIdentity::from_str(&identity)
-                    .map_err(|e| ConfigSetupError::ExitPoint(e.to_string()))?;
-                Ok(ExitPoint::Gateway {
-                    identity: node_identity,
-                })
-            }
-            ExitPointExtV2::Country {
-                two_letter_iso_country_code,
-            } => Ok(ExitPoint::Country {
-                two_letter_iso_country_code,
-            }),
-            ExitPointExtV2::Region { region } => Ok(ExitPoint::Region { region }),
-            ExitPointExtV2::Random => Ok(ExitPoint::Random),
-        }
-    }
-}
-
-impl TryFrom<&ExitPoint> for ExitPointExtV2 {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: &ExitPoint) -> Result<Self, Self::Error> {
-        match value {
-            ExitPoint::Address { address } => Ok(ExitPointExtV2::Address {
-                address: address.to_string(),
-            }),
-            ExitPoint::Gateway { identity } => Ok(ExitPointExtV2::Gateway {
-                identity: identity.to_string(),
-            }),
-            ExitPoint::Country {
-                two_letter_iso_country_code,
-            } => Ok(ExitPointExtV2::Country {
-                two_letter_iso_country_code: two_letter_iso_country_code.clone(),
-            }),
-            ExitPoint::Region { region } => Ok(ExitPointExtV2::Region {
-                region: region.clone(),
-            }),
-            ExitPoint::Random => Ok(ExitPointExtV2::Random),
-        }
-    }
-}
-
-//
-// EntryPointExtV1
-//
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-enum EntryPointExtV1 {
-    Gateway { identity: String },
-    Location { location: String },
-    Random,
-}
-
-impl TryFrom<EntryPointExtV1> for EntryPoint {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: EntryPointExtV1) -> Result<Self, Self::Error> {
-        match value {
-            EntryPointExtV1::Gateway { ref identity } => EntryPoint::from_base58_string(identity)
-                .map_err(|e| ConfigSetupError::EntryPoint(e.to_string())),
-            EntryPointExtV1::Location { location } => Ok(EntryPoint::Country {
-                two_letter_iso_country_code: location,
-            }),
-            EntryPointExtV1::Random => Ok(EntryPoint::Random),
-        }
-    }
-}
-
-//
-// ExitPointExtV1
-//
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "snake_case")]
-enum ExitPointExtV1 {
-    Address { address: String },
-    Gateway { identity: String },
-    Location { location: String },
-    Random,
-}
-
-impl TryFrom<ExitPointExtV1> for ExitPoint {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: ExitPointExtV1) -> Result<Self, Self::Error> {
-        match value {
-            ExitPointExtV1::Address { address } => {
-                let recipient = Recipient::try_from_base58_string(&address)
-                    .map_err(|e| ConfigSetupError::ExitPoint(e.to_string()))?;
-                Ok(ExitPoint::Address {
-                    address: Box::new(recipient),
-                })
-            }
-            ExitPointExtV1::Gateway { identity } => {
-                let node_identity = NodeIdentity::from_str(&identity)
-                    .map_err(|e| ConfigSetupError::ExitPoint(e.to_string()))?;
-                Ok(ExitPoint::Gateway {
-                    identity: node_identity,
-                })
-            }
-            ExitPointExtV1::Location { location } => Ok(ExitPoint::Country {
-                two_letter_iso_country_code: location,
-            }),
-            ExitPointExtV1::Random => Ok(ExitPoint::Random),
-        }
-    }
-}
-
-//
-// Legacy TOML version of config file.
-//
-
-#[derive(Clone, Debug, Deserialize)]
-enum LegacyEntryPoint {
-    Gateway { identity: Vec<u8> },
-    Location { location: String },
-    Random,
-}
-
-impl TryFrom<LegacyEntryPoint> for EntryPoint {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: LegacyEntryPoint) -> Result<Self, Self::Error> {
-        match value {
-            LegacyEntryPoint::Gateway { identity } => Ok(EntryPoint::Gateway {
-                identity: NodeIdentity::from_bytes(&identity)
-                    .map_err(|e| ConfigSetupError::EntryPoint(e.to_string()))?,
-            }),
-            LegacyEntryPoint::Location { location } => Ok(EntryPoint::Country {
-                two_letter_iso_country_code: location,
-            }),
-            LegacyEntryPoint::Random => Ok(EntryPoint::Random),
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize)]
-enum LegacyExitPoint {
-    Address { address: String },
-    Gateway { identity: Vec<u8> },
-    Location { location: String },
-    Random,
-}
-
-impl TryFrom<LegacyExitPoint> for ExitPoint {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: LegacyExitPoint) -> Result<Self, Self::Error> {
-        match value {
-            LegacyExitPoint::Address { address } => {
-                let recipient = Recipient::try_from_base58_string(&address)
-                    .map_err(|e| ConfigSetupError::ExitPoint(e.to_string()))?;
-                Ok(ExitPoint::Address {
-                    address: Box::new(recipient),
-                })
-            }
-            LegacyExitPoint::Gateway { identity } => Ok(ExitPoint::Gateway {
-                identity: NodeIdentity::from_bytes(&identity)
-                    .map_err(|e| ConfigSetupError::ExitPoint(e.to_string()))?,
-            }),
-            LegacyExitPoint::Location { location } => Ok(ExitPoint::Country {
-                two_letter_iso_country_code: location,
-            }),
-            LegacyExitPoint::Random => Ok(ExitPoint::Random),
-        }
-    }
-}
-
-#[derive(Clone, Debug, Deserialize)]
-#[serde(rename_all = "snake_case")]
-struct LegacyVpnServiceConfig {
-    entry_point: LegacyEntryPoint,
-    exit_point: LegacyExitPoint,
-}
-
-impl TryFrom<LegacyVpnServiceConfig> for VpnServiceConfig {
-    type Error = ConfigSetupError;
-
-    fn try_from(value: LegacyVpnServiceConfig) -> Result<Self, Self::Error> {
-        Ok(Self {
-            entry_point: value.entry_point.try_into()?,
-            exit_point: value.exit_point.try_into()?,
-            ..Default::default()
-        })
-    }
-}
-
-//
-// ConfigSetupError
-//
-
-#[derive(thiserror::Error, Debug)]
-pub enum ConfigSetupError {
-    #[error("failed to serialize JSON config file {file}")]
-    SerializeJson {
-        file: PathBuf,
-        #[source]
-        error: Box<serde_json::Error>,
-    },
-
-    #[error("failed to parse TOML config file {file}")]
-    ParseToml {
-        file: PathBuf,
-        #[source]
-        error: Box<toml::de::Error>,
-    },
-
-    #[error("failed to parse JSON config file {file}")]
-    ParseJson {
-        file: PathBuf,
-        #[source]
-        error: Box<serde_json::Error>,
-    },
-
-    #[error("failed to read config file {file}")]
-    ReadConfig {
-        file: PathBuf,
-        #[source]
-        error: io::Error,
-    },
-
-    #[error("failed to get parent directory of {file}")]
-    GetParentDirectory { file: PathBuf },
-
-    #[error("failed to create directory {dir}")]
-    CreateDirectory {
-        dir: PathBuf,
-        #[source]
-        error: io::Error,
-    },
-
-    #[error("failed to write file {file}")]
-    WriteFile { file: PathBuf, error: io::Error },
-
-    #[cfg(unix)]
-    #[error("failed to set permissions for directory {dir}")]
-    SetPermissions {
-        dir: PathBuf,
-        #[source]
-        error: io::Error,
-    },
-
-    #[cfg(windows)]
-    #[error("failed to set permissions for directory {dir}")]
-    SetPermissions {
-        dir: PathBuf,
-        #[source]
-        error: nym_windows::security::Error,
-    },
-
-    #[error("failed to convert entry point: {0}")]
-    EntryPoint(String),
-
-    #[error("failed to convert exit point: {0}")]
-    ExitPoint(String),
-
-    #[error("failed to convert IP address")]
-    IpAddress {
-        #[source]
-        error: Box<std::net::AddrParseError>,
-    },
-
-    #[error("failed to convert user agent {user_agent}")]
-    UserAgent {
-        user_agent: String, // Importing UserAgentError seems impossible.
-    },
-}
-
-#[cfg(windows)]
-pub fn program_data_path() -> PathBuf {
-    PathBuf::from(std::env::var("ProgramData").unwrap_or(std::env::var("PROGRAMDATA").unwrap()))
-}
-
-fn default_data_dir() -> PathBuf {
-    #[cfg(windows)]
-    return program_data_path().join("nym-vpnd").join("data");
-
-    #[cfg(not(windows))]
-    return DEFAULT_DATA_DIR.into();
-}
-
-pub fn data_dir() -> PathBuf {
-    std::env::var("NYM_VPND_DATA_DIR")
-        .map(PathBuf::from)
-        .unwrap_or_else(|_| default_data_dir())
-}
-
-fn default_log_dir() -> PathBuf {
-    #[cfg(windows)]
-    return program_data_path().join("nym-vpnd").join("log");
-
-    #[cfg(not(windows))]
-    return DEFAULT_LOG_DIR.into();
-}
-
-pub fn log_dir() -> PathBuf {
-    std::env::var("NYM_VPND_LOG_DIR")
-        .map(PathBuf::from)
-        .unwrap_or_else(|_| default_log_dir())
-}
-
-pub fn default_config_dir() -> PathBuf {
-    #[cfg(windows)]
-    return program_data_path().join("nym-vpnd").join("config");
-
-    #[cfg(not(windows))]
-    return DEFAULT_CONFIG_DIR.into();
-}
-
-pub fn config_dir() -> PathBuf {
-    std::env::var("NYM_VPND_CONFIG_DIR")
-        .map(PathBuf::from)
-        .unwrap_or_else(|_| default_config_dir())
-}
-
-pub async fn read_toml_config_file<C>(file_path: &Path) -> Result<C, ConfigSetupError>
-where
-    C: DeserializeOwned,
-{
-    let file_content =
-        fs::read_to_string(file_path)
-            .await
-            .map_err(|error| ConfigSetupError::ReadConfig {
-                file: file_path.to_path_buf(),
-                error,
-            })?;
-    toml::from_str(&file_content).map_err(|error| ConfigSetupError::ParseToml {
-        file: file_path.to_path_buf(),
-        error: Box::new(error),
-    })
-}
-
-pub async fn read_json_config_file<C>(file_path: &Path) -> Result<C, ConfigSetupError>
-where
-    C: DeserializeOwned,
-{
-    let bytes = tokio::fs::read(file_path)
-        .await
-        .map_err(|error| ConfigSetupError::ReadConfig {
-            file: file_path.to_path_buf(),
-            error,
-        })?;
-
-    serde_json::from_slice(&bytes).map_err(|error| ConfigSetupError::ParseJson {
-        file: file_path.to_path_buf(),
-        error: Box::new(error),
-    })
-}
-
-pub async fn write_json_config_file<C>(file_path: &Path, config: &C) -> Result<(), ConfigSetupError>
-where
-    C: Serialize,
-{
-    let json_bytes =
-        serde_json::to_vec_pretty(&config).map_err(|error| ConfigSetupError::SerializeJson {
-            file: file_path.to_path_buf(),
-            error: Box::new(error),
-        })?;
-
-    // Ensure parent directory exists
-    let config_dir = file_path
-        .parent()
-        .ok_or_else(|| ConfigSetupError::GetParentDirectory {
-            file: file_path.to_path_buf(),
-        })?;
-
-    fs::create_dir_all(config_dir)
-        .await
-        .map_err(|error| ConfigSetupError::CreateDirectory {
-            dir: config_dir.to_path_buf(),
-            error,
-        })?;
-
-    let file = fs::File::create(file_path)
-        .await
-        .map_err(|error| ConfigSetupError::WriteFile {
-            file: file_path.to_path_buf(),
-            error,
-        })?;
-
-    let mut writer = io::BufWriter::new(file);
-
-    writer
-        .write_all(&json_bytes)
-        .await
-        .map_err(|error| ConfigSetupError::WriteFile {
-            file: file_path.to_path_buf(),
-            error,
-        })?;
-    writer
-        .flush() // This is important!
-        .await
-        .map_err(|error| ConfigSetupError::WriteFile {
-            file: file_path.to_path_buf(),
-            error,
-        })
-}
-
-pub async fn create_data_dir(data_dir: &Path, network_name: &str) -> Result<(), ConfigSetupError> {
-    let network_data_dir = data_dir.join(network_name);
-
-    fs::create_dir_all(&network_data_dir)
-        .await
-        .map_err(|error| ConfigSetupError::CreateDirectory {
-            dir: network_data_dir.clone(),
-            error,
-        })?;
-
-    tracing::debug!(
-        "Making sure data dir exists at {}",
-        network_data_dir.display()
-    );
-
-    for dir_path in [&network_data_dir, data_dir] {
-        #[cfg(unix)]
-        {
-            // Set directory permissions to 700 (rwx------)
-            let permissions = std::fs::Permissions::from_mode(0o700);
-            fs::set_permissions(dir_path, permissions)
-                .await
-                .map_err(|error| ConfigSetupError::SetPermissions {
-                    dir: dir_path.to_path_buf(),
-                    error,
-                })?;
-        }
-
-        #[cfg(windows)]
-        {
-            set_data_dir_permissions(dir_path).map_err(|error| {
-                ConfigSetupError::SetPermissions {
-                    dir: dir_path.to_path_buf(),
-                    error,
-                }
-            })?;
-        }
-    }
-
-    Ok(())
-}
-
-/// Set directory permissions to Administrators with Full Control.
-#[cfg(windows)]
-fn set_data_dir_permissions(data_dir: &Path) -> nym_windows::security::Result<()> {
-    use nym_windows::security::{
-        AccessMode, AceFlags, Acl, ExplicitAccess, FileAccessRights, SecurityInfo,
-        SecurityObjectType, Sid, Trustee, TrusteeType, WellKnownSid, set_named_security_info,
-    };
-
-    let administrators_sid = Sid::well_known(WellKnownSid::BuiltinAdministrators)?;
-
-    let allow_admin_group_access = ExplicitAccess::new(
-        Trustee::new(administrators_sid.try_clone()?, TrusteeType::WellKnownGroup),
-        AccessMode::SetAccess,
-        FileAccessRights::FILE_ALL_ACCESS.into(),
-        AceFlags::OBJECT_INHERIT_ACE | AceFlags::CONTAINER_INHERIT_ACE,
-    );
-
-    let acl = Acl::new(vec![allow_admin_group_access])?;
-
-    set_named_security_info(
-        data_dir,
-        SecurityObjectType::FileObject,
-        SecurityInfo::DACL | SecurityInfo::PROTECTED_DACL,
-        None,
-        None,
-        Some(&acl),
-    )?;
-
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-    use tempfile::tempdir;
-
-    // Test migrating from TOML to the latest JSON version
-    async fn run_migrate_toml_test(
-        toml_content: &str,
-        json_latest_content: &str,
-        entry_point: nym_vpn_lib_types::EntryPoint,
-        exit_point: nym_vpn_lib_types::ExitPoint,
-    ) {
-        let temp_dir = tempdir().unwrap();
-        let config_path = temp_dir.path();
-
-        let network_config_path = config_path.join("tulips");
-        let _ = fs::create_dir_all(&network_config_path).await;
-        let toml_path = network_config_path.join(DEFAULT_CONFIG_FILE_TOML);
-        let json_path = network_config_path.join(DEFAULT_CONFIG_FILE_JSON);
-
-        // Write the TOML config file
-        fs::write(&toml_path, toml_content).await.unwrap();
-
-        // Read the TOML config and migrate it to latest JSON version.  The latest version of the
-        // JSON will be written straight back to disk.
-        let config_manager = VpnServiceConfigManager::new(&network_config_path, None)
-            .await
-            .unwrap();
-        let config = config_manager.config();
-        assert_eq!(config.entry_point, entry_point);
-        assert_eq!(config.exit_point, exit_point);
-
-        // The TOML file should be deleted and replaced with a JSON version
-        assert!(!toml_path.exists());
-        assert!(json_path.exists());
-
-        // Read the JSON config
-        let config_manager = VpnServiceConfigManager::new(&network_config_path, None)
-            .await
-            .unwrap();
-        let config = config_manager.config();
-        assert_eq!(config.entry_point, entry_point);
-        assert_eq!(config.exit_point, exit_point);
-
-        // Check the JSON is the right version and all snake-case
-        let read_json_content = fs::read_to_string(&json_path).await.unwrap();
-        assert_eq!(json_latest_content, read_json_content);
-    }
-
-    // Test migrating from an old JSON version to the latest JSON version
-    async fn run_migrate_json_test(json_old_content: &str, json_latest_content: &str) {
-        let temp_dir = tempdir().unwrap();
-        let config_path = temp_dir.path();
-
-        let network_config_path = config_path.join("tulips");
-        let _ = fs::create_dir_all(&network_config_path).await;
-        let json_path = network_config_path.join(DEFAULT_CONFIG_FILE_JSON);
-
-        // Write the old JSON config file
-        fs::write(&json_path, json_old_content).await.unwrap();
-
-        // Read the old JSON config and migrate it to latest JSON.  The latest version of the
-        // JSON will be written straight back to disk.
-        let _config_manager = VpnServiceConfigManager::new(&network_config_path, None)
-            .await
-            .unwrap();
-
-        // Check the JSON is the right version and all snake-case (ignore whitespace/order)
-        let read_json_content = fs::read_to_string(&json_path).await.unwrap();
-        let expected: serde_json::Value = serde_json::from_str(json_latest_content).unwrap();
-        let actual: serde_json::Value = serde_json::from_str(&read_json_content).unwrap();
-        assert_eq!(expected, actual);
-    }
-
-    // Test serializing and deserializing the config produces the same result
-    async fn run_serialize_test(config: VpnServiceConfig) {
-        let temp_dir = tempdir().unwrap();
-        let config_path = temp_dir.path();
-
-        let network_config_path = config_path.join("tulips");
-
-        // Write the config to disk
-        let mut config_manager = VpnServiceConfigManager::new(&network_config_path, None)
-            .await
-            .unwrap();
-        config_manager.set_config(config.clone()).await;
-        assert!(config_manager.write_to_file().await);
-        drop(config_manager);
-
-        // Read it back and compare it
-        let config_manager = VpnServiceConfigManager::new(&network_config_path, None)
-            .await
-            .unwrap();
-        let read_config = config_manager.config();
-        assert_eq!(&config, read_config);
-    }
-
-    // Test reading a broken config falls back to a default config
-    async fn run_fallback_test(broken_json_content: &str) {
-        let temp_dir = tempdir().unwrap();
-        let config_path = temp_dir.path();
-
-        let network_config_path = config_path.join("tulips");
-        let _ = fs::create_dir_all(&network_config_path).await;
-        let json_path = network_config_path.join(DEFAULT_CONFIG_FILE_JSON);
-
-        // Write the broken JSON config file
-        fs::write(&json_path, broken_json_content).await.unwrap();
-
-        // Ensure reading the broken config falls back to default config
-        let config_manager = VpnServiceConfigManager::new(&network_config_path, None)
-            .await
-            .unwrap();
-
-        assert_eq!(config_manager.config(), &VpnServiceConfig::default());
-    }
-
-    #[tokio::test]
-    async fn test_service_config_migrate_toml_location() {
-        let toml_content = r#"
-[entry_point.Location]
-location = "FR"
-
-[exit_point.Location]
-location = "BE"
-"#;
-
-        let json_content = r#"{
-  "version": "v2",
-  "entry_point": {
-    "country": {
-      "two_letter_iso_country_code": "FR"
-    }
-  },
-  "exit_point": {
-    "country": {
-      "two_letter_iso_country_code": "BE"
-    }
-  },
-  "dns": null,
-  "allow_lan": false,
-  "disable_ipv6": false,
-  "enable_two_hop": false,
-  "enable_bridges": false,
-  "netstack": false,
-  "disable_poisson_rate": false,
-  "disable_background_cover_traffic": false,
-  "min_mixnode_performance": null,
-  "min_gateway_mixnet_performance": null,
-  "min_gateway_vpn_performance": null,
-  "residential_exit": false
-}"#;
-
-        let entry_point = nym_vpn_lib_types::EntryPoint::Country {
-            two_letter_iso_country_code: "FR".to_string(),
-        };
-
-        let exit_point = nym_vpn_lib_types::ExitPoint::Country {
-            two_letter_iso_country_code: "BE".to_string(),
-        };
-
-        run_migrate_toml_test(toml_content, json_content, entry_point, exit_point).await;
-    }
-
-    #[tokio::test]
-    async fn test_service_config_migrate_gateway() {
-        let toml_content = r#"
-[entry_point.Gateway]
-identity = [ 92, 25, 33, 77, 4, 117, 82, 117, 246, 239, 233, 11, 129, 183, 86, 194, 140, 95, 21, 196, 121, 130, 232, 195, 71, 173, 66, 124, 5, 14, 114, 107, ]
-
-[exit_point.Gateway]
-identity = [ 99, 23, 98, 234, 66, 161, 195, 63, 155, 161, 250, 207, 17, 158, 136, 114, 215, 90, 236, 161, 231, 176, 140, 190, 147, 182, 64, 171, 145, 31, 245, 186, ]
-"#;
-
-        let json_content = r#"{
-  "version": "v2",
-  "entry_point": {
-    "gateway": {
-      "identity": "7CWjY3QFoA9dgE535u9bQiXCfzgMZvSpJu842GA1Wn42"
-    }
-  },
-  "exit_point": {
-    "gateway": {
-      "identity": "7fp3cmzCvgeRgbB1ycTnK6RokjHNqPmCCSBG23gyxshj"
-    }
-  },
-  "dns": null,
-  "allow_lan": false,
-  "disable_ipv6": false,
-  "enable_two_hop": false,
-  "enable_bridges": false,
-  "netstack": false,
-  "disable_poisson_rate": false,
-  "disable_background_cover_traffic": false,
-  "min_mixnode_performance": null,
-  "min_gateway_mixnet_performance": null,
-  "min_gateway_vpn_performance": null,
-  "residential_exit": false
-}"#;
-
-        let entry_point = nym_vpn_lib_types::EntryPoint::Gateway {
-            identity: nym_vpn_lib_types::NodeIdentity::from_str(
-                "7CWjY3QFoA9dgE535u9bQiXCfzgMZvSpJu842GA1Wn42",
-            )
-            .unwrap(),
-        };
-
-        let exit_point = nym_vpn_lib_types::ExitPoint::Gateway {
-            identity: nym_vpn_lib_types::NodeIdentity::from_str(
-                "7fp3cmzCvgeRgbB1ycTnK6RokjHNqPmCCSBG23gyxshj",
-            )
-            .unwrap(),
-        };
-
-        run_migrate_toml_test(toml_content, json_content, entry_point, exit_point).await;
-    }
-
-    #[tokio::test]
-    #[ignore] // Temporarily disabled due to issues with ExitPoint::Address (de)serialisation.
-    async fn test_service_config_migrate_toml_address() {
-        let toml_content = r#"
-[entry_point.Gateway]
-identity = [92, 25, 33, 77, 4, 117, 82, 117, 246, 239, 233, 11, 129, 183, 86, 194, 140, 95, 21, 196, 121, 130, 232, 195, 71, 173, 66, 124, 5, 14, 114, 107]
-
-[exit_point.Address]
-address = [5, 56, 84, 195, 94, 238, 210, 124, 65, 143, 209, 144, 22, 255, 91, 188, 35, 50, 144, 234, 226, 114, 99, 40, 10, 102, 200, 170, 19, 162, 86, 134, 84, 20, 195, 193, 42, 194, 230, 153, 163, 90, 214, 216, 196, 166, 87, 132, 206, 215, 91, 89, 51, 98, 72, 156, 159, 248, 109, 225, 152, 204, 80, 97, 9, 62, 22, 108, 155, 95, 153, 29, 143, 48, 208, 5, 101, 231, 176, 93, 107, 229, 11, 225, 145, 1, 14, 219, 44, 88, 199, 206, 40, 185, 150, 151]
-"#;
-
-        let json_content = r#"{
-  "version": "v2",
-  "entry_point": {
-    "gateway": {
-      "identity": "7CWjY3QFoA9dgE535u9bQiXCfzgMZvSpJu842GA1Wn42"
-    }
-  },
-  "exit_point": {
-    "address": {
-            "address": "MNrmKzuKjNdbEhfPUzVNfjw63oBQNSayqoQKGL4JjAV.6fDcSN6faGpvA3pd3riCwjpzXc7RQfWmGMa82UVoEwKE@d5adfJNtcdZW2XwK85JAAU8nXAs9JCPYn2RNvDLZn4e"
-    }
-  },
-  "dns": null,
-  "disable_ipv6": false,
-  "enable_two_hop": false,
-  "enable_bridges": false,
-  "netstack": false,
-  "disable_poisson_rate": false,
-  "disable_background_cover_traffic": false,
-  "min_mixnode_performance": null,
-  "min_gateway_mixnet_performance": null,
-  "min_gateway_vpn_performance": null,
-  "residential_exit": false
-}"#;
-
-        let entry_point = nym_vpn_lib_types::EntryPoint::Gateway {
-            identity: nym_vpn_lib_types::NodeIdentity::from_str(
-                "7CWjY3QFoA9dgE535u9bQiXCfzgMZvSpJu842GA1Wn42",
-            )
-            .unwrap(),
-        };
-
-        let exit_point = nym_vpn_lib_types::ExitPoint::Address {
-            address: Box::new(
-                nym_vpn_lib_types::Recipient::from_str("MNrmKzuKjNdbEhfPUzVNfjw63oBQNSayqoQKGL4JjAV.6fDcSN6faGpvA3pd3riCwjpzXc7RQfWmGMa82UVoEwKE@d5adfJNtcdZW2XwK85JAAU8nXAs9JCPYn2RNvDLZn4e").unwrap(),
-            )
-        };
-
-        run_migrate_toml_test(toml_content, json_content, entry_point, exit_point).await;
-    }
-
-    #[tokio::test]
-    async fn test_service_config_migrate_toml_random() {
-        let toml_content = r#"
-entry_point = "Random"
-exit_point = "Random"
-"#;
-
-        let json_content = r#"{
-  "version": "v2",
-  "entry_point": "random",
-  "exit_point": "random",
-  "dns": null,
-  "allow_lan": false,
-  "disable_ipv6": false,
-  "enable_two_hop": false,
-  "enable_bridges": false,
-  "netstack": false,
-  "disable_poisson_rate": false,
-  "disable_background_cover_traffic": false,
-  "min_mixnode_performance": null,
-  "min_gateway_mixnet_performance": null,
-  "min_gateway_vpn_performance": null,
-  "residential_exit": false
-}"#;
-
-        let entry_point = nym_vpn_lib_types::EntryPoint::Random;
-
-        let exit_point = nym_vpn_lib_types::ExitPoint::Random;
-
-        run_migrate_toml_test(toml_content, json_content, entry_point, exit_point).await;
-    }
-
-    #[tokio::test]
-    async fn test_service_config_migrate_from_v1() {
-        let json_v1_content = r#"{
-  "version": "v1",
-  "entry_point": {
-    "gateway": {
-      "identity": "7CWjY3QFoA9dgE535u9bQiXCfzgMZvSpJu842GA1Wn42"
-    }
-  },
-  "exit_point": {
-    "address": {
-      "address": "MNrmKzuKjNdbEhfPUzVNfjw63oBQNSayqoQKGL4JjAV.6fDcSN6faGpvA3pd3riCwjpzXc7RQfWmGMa82UVoEwKE@d5adfJNtcdZW2XwK85JAAU8nXAs9JCPYn2RNvDLZn4e"
-    }
-  }
-}"#;
-
-        let json_latest_content = r#"{
-  "version": "v2",
-  "entry_point": {
-    "gateway": {
-      "identity": "7CWjY3QFoA9dgE535u9bQiXCfzgMZvSpJu842GA1Wn42"
-    }
-  },
-  "exit_point": {
-    "address": {
-      "address": "MNrmKzuKjNdbEhfPUzVNfjw63oBQNSayqoQKGL4JjAV.6fDcSN6faGpvA3pd3riCwjpzXc7RQfWmGMa82UVoEwKE@d5adfJNtcdZW2XwK85JAAU8nXAs9JCPYn2RNvDLZn4e"
-    }
-  },
-  "dns": null,
-  "allow_lan": false,
-  "disable_ipv6": false,
-  "enable_two_hop": false,
-  "enable_bridges": false,
-  "netstack": false,
-  "disable_poisson_rate": false,
-  "disable_background_cover_traffic": false,
-  "min_mixnode_performance": null,
-  "min_gateway_mixnet_performance": null,
-  "min_gateway_vpn_performance": null,
-  "residential_exit": false
-}"#;
-
-        run_migrate_json_test(json_v1_content, json_latest_content).await;
-    }
-
-    #[tokio::test]
-    async fn test_service_config_fallback_default_v1() {
-        let broken_json_content = r#"{
-  "version": "v1",
-  "entrXy_point": {
-    "gateway": {
-      "identity": "7CWjY3QFoA9dgE535u9bQiXCfzgMZvSpJu842GA1Wn42"
-    }
-  },
-  "exit_point": {
-    "address": {
-      "address": "MNrmKzuKjNdbEhfPUzVNfjw63oBQNSayqoQKGL4JjAV.6fDcSN6faGpvA3pd3riCwjpzXc7RQfWmGMa82UVoEwKE@d5adfJNtcdZW2XwK85JAAU8nXAs9JCPYn2RNvDLZn4e"
-    }
-  }
-}"#;
-        run_fallback_test(broken_json_content).await;
-    }
-
-    #[tokio::test]
-    async fn test_service_config_fallback_default_v2() {
-        let broken_json_content = r#"{
-  "version": "v2",
-  "entry_point": {
-    "gateway": {
-      "identity": "7CWjY3QFoA9dgE535u9bQiXCfzgMZvSpJu842GA1Wn42"
-    }
-  },
-  "exit_point": {
-    "address": {
-      "address": "MNrmKzuKjNdbEhfPUzVNfjw63oBQNSayqoQKGL4JjAV.6fDcSN6faGpvA3pd3riCwjpzXc7RQfWmGMa82UVoEwKE@d5adfJNtcdZW2XwK85JAAU8nXAs9JCPYn2RNvDLZn4e"
-    }
-  },
-  "dns": null,
-  "disable_ipv6": false,
-  "enable_two_hop": false,
-  "enable_bridges": false,
-  "netstack": false,
-  "disable_poisson_rate": false,
-  "disable_background_cover_traffic": false,
-  "min_mixnode_performance": null,
-  "min_gateway_mixnet_performance": null,
-  "min_gateway_vpn_performance": null,
-  "residential_exit": false
-}"#;
-
-        run_fallback_test(broken_json_content).await;
-    }
-
-    #[tokio::test]
-    async fn test_service_config_serialize_defaults() {
-        let config = VpnServiceConfig::default();
-        run_serialize_test(config).await;
-    }
-
-    #[tokio::test]
-    async fn test_service_config_serialize_full() {
-        let config = VpnServiceConfig {
-            entry_point: nym_vpn_lib_types::EntryPoint::Country {
-                two_letter_iso_country_code: "US".to_string(),
-            },
-            exit_point: nym_vpn_lib_types::ExitPoint::Gateway {
-                identity: nym_vpn_lib_types::NodeIdentity::from_str(
-                    "7fp3cmzCvgeRgbB1ycTnK6RokjHNqPmCCSBG23gyxshj",
-                )
-                .unwrap(),
-            },
-            dns: Some(IpAddr::from_str("192.168.50.1").unwrap()),
-            allow_lan: true,
-            disable_ipv6: true,
-            enable_two_hop: true,
-            enable_bridges: false,
-            netstack: true,
-            disable_poisson_rate: true,
-            disable_background_cover_traffic: true,
-            min_mixnode_performance: Some(55u8),
-            min_gateway_mixnet_performance: Some(64u8),
-            min_gateway_vpn_performance: Some(1u8),
-            residential_exit: true,
-        };
-        run_serialize_test(config).await;
-    }
-}
